fix: use layout.svelte
